requirements.txt
# PCB 瑕疵偵測系統 - 環境依賴套件
# Python >= 3.9 建議

# 核心深度學習框架
torch>=2.0.0
torchvision>=0.15.0

# YOLOv8 (Ultralytics)
ultralytics>=8.0.0

# 影像處理
opencv-python>=4.8.0
Pillow>=10.0.0
albumentations>=1.3.0

# 資料處理
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0

# GUI 介面
gradio>=4.0.0

# 資料集下載與管理
gdown>=4.7.0
roboflow>=1.1.0

# 工具套件
tqdm>=4.65.0
pyyaml>=6.0
python-dotenv>=1.0.0

# 報告與視覺化
tensorboard>=2.13.0
